% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alb5_rev.R
\docType{data}
\name{alb5_rev}
\alias{alb5_rev}
\title{Revision protocol dataframe of replication project ML5
Study: Alb5}
\format{
A data frame with 884 rows and 4 variables:
\describe{
  \item{ResponseId}{unique ID of participant}
  \item{Location}{Grouping variable for where the replication of the experiment took place}
  \item{SATTotal}{Sum of all SAT variables, and dependent variable in our experiment}
  \item{Condition}{Experimental condition of the experiment}
  ...
}
}
\source{
\url{https://osf.io/64njs/}
}
\usage{
alb5_rev
}
\description{
Revision protocol dataframe of replication project ML5
Study: Alb5
}
\keyword{datasets}
