% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{CI_FE_moderation}
\alias{CI_FE_moderation}
\title{Moderation analysis. We construct the Wlad CI for the fixed effects. If any interaction term's coefficient does 
cover zero, we keep the term in the model}
\usage{
CI_FE_moderation(fitModer)
}
\arguments{
\item{ModelCorr}{it is the output from the function LRTfunction (we need the fit of the "winning model")}
}
\value{
Which model is chosen? p-values supporting the decision
}
\description{
Moderation analysis. We construct the Wlad CI for the fixed effects. If any interaction term's coefficient does 
cover zero, we keep the term in the model
}
\author{
Federico Rogai
}
