% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{diagnosticPlot}
\alias{diagnosticPlot}
\title{Diagnostic Plot for "winning model" as defined by the likelihood ratio test.}
\usage{
diagnosticPlot(winningModel, percent = 1)
}
\arguments{
\item{winningModel}{best model, according to LRT}

\item{percent}{used to avoid overplotting in Tukey-Anscombe plot.}
}
\value{
Tukey-Anscombe plot, Q-Q plot for RE and residuals' Q-Q plot.
}
\description{
Diagnostic Plot for "winning model" as defined by the likelihood ratio test.
}
\author{
Federico Rogai
}
