% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{diagnosticPlot.study}
\alias{diagnosticPlot.study}
\title{Diagnostic Plot for "winning model" as defined by the likelihood ratio test.}
\usage{
diagnosticPlot.study(object, percent = 1)
}
\arguments{
\item{object}{study object}

\item{percent}{used to avoid overplotting in Tukey-Anscombe plot.}
}
\value{
Tukey-Anscombe plot, Q-Q plot for RE and residuals' Q-Q plot.
}
\description{
Diagnostic Plot for "winning model" as defined by the likelihood ratio test.
}
\author{
Stefan Thoma, adapted from Federico Rogai
}
