% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_study-obj.R
\docType{data}
\name{digitsForRounding}
\alias{digitsForRounding}
\title{Define the constant which will be used to round digits.}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
digitsForRounding
}
\description{
Define the constant which will be used to round digits.
}
\keyword{datasets}
