% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtmturk.R
\docType{data}
\name{dtmturk}
\alias{dtmturk}
\title{Replication protocol dataframe of replication project ML5
Study: Alb5}
\format{
A data frame with 580 rows and 33 variables:
\describe{
  \item{ResponseId}{unique ID of participant}
  \item{Location}{Grouping variable for where the replication of the experiment took place. In this case just mTurk}
  \item{SATTotal}{Sum of all SAT variables, and dependent variable in our experiment}
  \item{Condition}{Experimental condition of the experiment}
  ...
}
}
\source{
\url{https://osf.io/z2fu9/}
}
\usage{
dtmturk
}
\description{
Replication protocol dataframe of replication project ML5
Study: Alb5
}
\keyword{datasets}
