% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-fn.R
\name{effect_differences.study}
\alias{effect_differences.study}
\title{WIP (not standardised yet)}
\usage{
effect_differences.study(object, coverage_prob = 0.95, standardise = NULL)
}
\arguments{
\item{object}{Study object.}

\item{coverage_prob}{Double. Coverage probability of the CIs. Defaults to .95}
}
\value{
CI. numeric vector containing the difference in effect sizes and the
lower and upper boundaries of a confidence interval for this difference.
}
\description{
WIP (not standardised yet)
}
\author{
Stefan Thoma, adapted from Lorenz Herger
}
