% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{interactPlot}
\alias{interactPlot}
\title{Interaction plot for one Research Question}
\usage{
interactPlot(variables, dat)
}
\arguments{
\item{variables}{one element of the list use_var_list, containing response and explanatory variable}

\item{dat}{dataset used}
}
\value{
Interaction plot
}
\description{
Interaction plot for one Research Question
}
\author{
Federico Rogai
}
