% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{LmerMA}
\alias{LmerMA}
\title{Moderation analysis for lmer objects. We compute an F-test in anove (SS of type III) and use the p-values
computed by the package lmerTest to assess the significance of the interaction terms. Unfortunately such an 
approach cannot be used for glmer objects (lmerTest does not return p-values for this method). 
For the sake of consistency we use the function CI_FE_moderation to do that.
FUNCTION NOT USED IN THE END}
\usage{
LmerMA(fitModer)
}
\arguments{
\item{ModelCorr}{it is the output from the function LRTfunction (we need the fit of the "winning model")}
}
\value{
Which model is chosen? p-values supporting the decision
}
\description{
Moderation analysis for lmer objects. We compute an F-test in anove (SS of type III) and use the p-values
computed by the package lmerTest to assess the significance of the interaction terms. Unfortunately such an 
approach cannot be used for glmer objects (lmerTest does not return p-values for this method). 
For the sake of consistency we use the function CI_FE_moderation to do that.
FUNCTION NOT USED IN THE END
}
\author{
Federico Rogai
}
