% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{meta_analysis}
\alias{meta_analysis}
\title{Classical Meta-Analysis procedure}
\usage{
meta_analysis(input_table, level = 0.95, calculate_OR = FALSE)
}
\arguments{
\item{input_table}{single element of the list effect_tables (contain effect's estimate and sd associated with it)}
}
\value{
Overall estimated effect and CI for it. Moreover Q-value and associated p-values are reported
to assess heterogeneity in the estimated effects
}
\description{
Classical Meta-Analysis procedure
}
\author{
Federico Rogai
}
