% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{MixedModels.study}
\alias{MixedModels.study}
\title{Function which fits different mixed models, compares them and finally defines a random
 coefficients model which, if neccessary, is reduced in random effects to minimize
 model singularity.
 This function also computes and returns model ICC, a measure of heterogeneity.}
\usage{
MixedModels.study(object, methodFamily = NULL, transformation = NULL)
}
\arguments{
\item{object}{of class study.}

\item{methodFamily}{which family does the error distribution belong to?
not required, can be used to overwrite the default defined in the object.
Defaults to NULL.}

\item{transformation}{name of the transformation that should be applied to the response (e.g. log).
defaults to NULL.}
}
\value{
return.list contains three fitted models, the information whether the random coef. model is singular, and the ICC table.
}
\description{
Function which fits different mixed models, compares them and finally defines a random
 coefficients model which, if neccessary, is reduced in random effects to minimize
 model singularity.
 This function also computes and returns model ICC, a measure of heterogeneity.
}
\author{
Stefan Thoma, adapted from Federico Rogai
}
