% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml1.R
\docType{data}
\name{ml1}
\alias{ml1}
\title{Revision protocol dataframe of replication project ML1
All studies.}
\format{
A data frame with 6344 rows and 382 variables:
}
\source{
\url{}
}
\usage{
ml1
}
\description{
See ... for data description
}
\keyword{datasets}
