% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{moderationAnalysis}
\alias{moderationAnalysis}
\title{This function inserts the fixed effects predictors "us_or_international" and "lab_or_online" in the "winning 
 model" from the Likelihood ratio test. These two predictors are separately made interact with the treatment
 effect to evaluate if they are to be considered moderators.}
\usage{
moderationAnalysis(ModelCorr)
}
\arguments{
\item{ModelCorr}{it is the output from the function LRTfunction (we need the fit of the "winning model")}
}
\value{
See CI_FE_moderation
}
\description{
This function inserts the fixed effects predictors "us_or_international" and "lab_or_online" in the "winning 
 model" from the Likelihood ratio test. These two predictors are separately made interact with the treatment
 effect to evaluate if they are to be considered moderators.
}
\author{
Federico Rogai
}
