% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{moderationInteractionPlot}
\alias{moderationInteractionPlot}
\title{Interaction plot for one Research Question (within colored lines for moderator)}
\usage{
moderationInteractionPlot(variables, dat, moderator)
}
\arguments{
\item{variables}{one element of the list use_var_list, containing response and explanatory variable}

\item{dat}{dataset used}
}
\value{
Interaction plot
}
\description{
Interaction plot for one Research Question (within colored lines for moderator)
}
\author{
Federico Rogai
}
