% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\docType{data}
\name{object}
\alias{object}
\title{Function which fits two mixed models and a fixed effect model and returns objects containing them.}
\format{
An object of class \code{study} of length 1.
}
\usage{
object
}
\arguments{
\item{object}{takes an object of class study as input}

\item{methodFamily}{which family does the error distribution belong to?
not required, can be used to overwrite the default defined in the object.}

\item{transformation}{name of the transformation that should be applied to the response (e.g. log).}
}
\value{
list with three objects of fitted models. The first is the "empty model", i.e. a model without random effects
but still with the fixed treatment effect. The second is the model with the random interaction. The third is the model
with an interaction between treatment and (random) university effect.
}
\description{
Function which fits two mixed models and a fixed effect model and returns objects containing them.
}
\author{
Stefan Thoma, adapted from Federico Rogai
}
\keyword{datasets}
