% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payne.R
\docType{data}
\name{payne}
\alias{payne}
\title{Dataframe of replication project ML5
Study: Payne (Work in progress)}
\format{
A data frame with 1103 rows and 15 variables:
\describe{
  \item{ResponseId}{unique ID of participant}
  \item{Location}{Grouping variable for where the replication of the experiment took place. In this case just mTurk}
  ...
}
}
\source{
\url{}
}
\usage{
payne
}
\description{
Dataframe of replication project ML5
Study: Payne (Work in progress)
}
\keyword{datasets}
