% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{plot_both.study}
\alias{plot_both.study}
\title{This function plots the relevance_overview_table and the difference table}
\usage{
plot_both.study(
  object,
  standardise = FALSE,
  coverage_probability = 0.95,
  cutoff.est = NULL,
  cutoff.diff = NULL
)
}
\arguments{
\item{object}{a study object}

\item{coverage_probability}{Double. Coverage probability of the CIs. Defaults}

\item{cutoff.est}{is forwarded as cutoff for estimate_plot function}

\item{cutoff.diff}{is forwarded as cutoff for difference_plot function
to .95}

\item{standardize}{should plot be based on standardized values.}
}
\description{
This function plots the relevance_overview_table and the difference table
}
\author{
Stefan Thoma
}
