% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{plot_difference.study}
\alias{plot_difference.study}
\title{This function plots the difference table as a forest plot
Adapted from https://www.selfmindsociety.com/post/a-forest-plot-in-ggplot2}
\usage{
plot_difference.study(
  object,
  cutoff = NULL,
  ci.type = "wald",
  standardise = TRUE
)
}
\arguments{
\item{object}{a study object}

\item{ci.type}{either "wald" or "newcombe"
to .95}

\item{coverage_probability}{Coverage probability of the CIs. Defaults}
}
\value{
plot
}
\description{
This function plots the difference table as a forest plot
Adapted from https://www.selfmindsociety.com/post/a-forest-plot-in-ggplot2
}
\author{
Stefan Thoma
}
