% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{plot_estimates.study}
\alias{plot_estimates.study}
\title{This function plots the table created by the function relevance_table.study}
\usage{
plot_estimates.study(object, cutoff = NULL, standardise = TRUE)
}
\arguments{
\item{object}{a study object}

\item{cutoff}{values used to crop the plot. Can be either one value or two.}

\item{standardise}{should standardised values be plotted?}
}
\value{
ggplot of relevance table
}
\description{
This function plots the table created by the function relevance_table.study
}
\author{
Stefan Thoma
}
