% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{power.f}
\alias{power.f}
\title{Calculates Power given a relvance table object.}
\usage{
power.f(x, std.effect, threshold, family)
}
\arguments{
\item{x}{predictor values. either vector or matrxi.
not sure if matrix works for all cases.}

\item{std.effect}{is the standardised effect. Only binom regression allows for intercept}

\item{threshold}{relevance threshold}

\item{family}{either gaussian or binomial}
}
\value{
power estimate based on 1000 simulation
}
\description{
Calculates Power given a relvance table object.
}
\author{
Stefan Thoma
}
