% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{relevance_f.study}
\alias{relevance_f.study}
\title{Creates a table showing the standardized mean difference obtained by
each of the 36 universities that tried to replicate a specific study.
This is an alternative function to escalc_inputs_smd and is based on the
relevance package by Stahel. The function is called by the relevance_table.study function}
\usage{
relevance_f.study(
  dat,
  treatment_var,
  response_var,
  orig,
  var.of.interest,
  family = "gaussian",
  relevance.threshold = 0.1
)
}
\arguments{
\item{dat}{data.frame containing the full data from the manyLabs experiments}

\item{treatment_var}{Name of the column in dat that contains the treatment
assignments.}

\item{response_var}{Name of the column in dat that contains the response.}

\item{orig}{Statistic of original study}
}
\value{
output data.frame with one row for each of the 36 replication attempts.
Each row of output contains the sample sizes and means for both treatment groups
as well as the the standardized mean difference.
}
\description{
Creates a table showing the standardized mean difference obtained by
each of the 36 universities that tried to replicate a specific study.
This is an alternative function to escalc_inputs_smd and is based on the
relevance package by Stahel. The function is called by the relevance_table.study function
}
\author{
Stefan Thoma
}
