% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{relevance_table.study}
\alias{relevance_table.study}
\title{creates a table with the effectsizes of all locations for a specific
 research question and one-sided CIs for
 the effect found by each university and the prediction interval of the overall effect.
 If the measure is either OR or SMD, it calls the function relevance_f.study to do most
 of the heavy lifting.
 If measure is drop, the function r_squared.study is called and its output returned.}
\usage{
relevance_table.study(
  object,
  coverage_prob = 0.95,
  mem = TRUE,
  use.both = TRUE
)
}
\arguments{
\item{object}{study object}

\item{mem}{boolean, whether fixed effect from mixed effects model should be included}

\item{use.both}{boolean, should both data frames be used?}

\item{coverage_probab}{Double. Coverage probability of the CIs. Defaults
to .95}
}
\value{
after_esclalc data.frame. Contains the university-level effect sizes
and CIs for one research question.
}
\description{
creates a table with the effectsizes of all locations for a specific
 research question and one-sided CIs for
 the effect found by each university and the prediction interval of the overall effect.
 If the measure is either OR or SMD, it calls the function relevance_f.study to do most
 of the heavy lifting.
 If measure is drop, the function r_squared.study is called and its output returned.
}
\author{
Stefan Thoma, adapted from Lorenz Herger
}
