% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales}
\alias{scales}
\title{Revision protocol dataframe of replication project ML1
All studies.}
\format{
A data frame with 1149 rows and 4 variables.
}
\source{
\url{}
}
\usage{
scales
}
\description{
See ... for data description
The data has been reduced to only 10 locations.
}
\keyword{datasets}
