% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{stnd.beta.glmer}
\alias{stnd.beta.glmer}
\title{Standardise Effects of lmer-model
This function is used by the create_overview_table function from replication-fn.R}
\usage{
stnd.beta.glmer(mod)
}
\arguments{
\item{lmer}{best model, according to LRT}
}
\value{
standardised effects + standardised CI based on
}
\description{
Standardise Effects of lmer-model
This function is used by the create_overview_table function from replication-fn.R
}
\author{
Stefan Thoma
}
