% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_study-obj.R
\docType{class}
\name{study-class}
\alias{study-class}
\alias{study}
\title{An S4 class to represent a many-labs replication study.}
\description{
An S4 class to represent a many-labs replication study.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of study in question}

\item{\code{manyLabs}}{From which many-labs project is the data (so far only ML1 & ML5 are supported)}

\item{\code{data.revised}}{Data generated with revised protocol (ML5)}

\item{\code{data.replication}}{Data generated with replication protocol (ML1 & ML5)}

\item{\code{original}}{Parameter of the original study. Best to use output of relevance::inference function and numer ob observations}

\item{\code{variables}}{X, Y, and type of Measure (OR, SMD or Drop1)}

\item{\code{var.of.interest}}{Depends on the Measure. Usually just one predictorn.}

\item{\code{relevance.threshold}}{Supply study specific relevance threshold}

\item{\code{family}}{Model family forwarded to glm and glmer}

\item{\code{mixedModels}}{Here you save the output of the MixedModels.study function}

\item{\code{table}}{Here you save the output of the relevance_table.study function}

\item{\code{difference.table}}{Here you save the output of the effect_differences.study function}
}}

