% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{success}
\alias{success}
\title{This function takes as input the output of relevance_overview_tables
It creates a new column that indicates whether a replication has been successful or not.}
\usage{
success(table, threshold = NULL)
}
\arguments{
\item{table}{list containing the full data and other infos from the manyLabs experiment
to .95}

\item{threshold}{relevance threshold. Needed if Rlp & Rls are not in table.}
}
\value{
vector indicating whether repl. has been successful or not.
}
\description{
This function takes as input the output of relevance_overview_tables
It creates a new column that indicates whether a replication has been successful or not.
}
\author{
Stefan Thoma
}
