% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{summary.study}
\alias{summary.study}
\title{summary.study, function to summarise the data frames of the study objects.
At the moment, this funcion only works for the schwarz (scales) and the alb5 study.}
\usage{
summary.study(object)
}
\arguments{
\item{object}{Object of class study}
}
\value{
summary measures about the data, depending on the measure specified.
}
\description{
summary.study, function to summarise the data frames of the study objects.
At the moment, this funcion only works for the schwarz (scales) and the alb5 study.
}
\author{
Stefan Thoma
}
