% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heterogen-fn.R
\name{TestCorrRE}
\alias{TestCorrRE}
\title{This function tests if a model in which the correlation between the two random effects (random intercept and random slope)
 is removed fits the data as good as one in which this term is present. If it does, it is compared to the random intercept 
 model only.}
\usage{
TestCorrRE(ModelCorr)
}
\arguments{
\item{ModelCorr}{it is the output from the function MixedModels. 
It is passed immediately to LRT since we need the fit of the "winning model".}
}
\value{
Returns an output from anova(model.1, model.2), which tells us
a. Is the model without correlation as good as the one in which we allow for it?
b. if the answer for a) is "Yes", we use a LRT as implemented in anova() to compare this model to the one
with the random Intercept only
}
\description{
This function tests if a model in which the correlation between the two random effects (random intercept and random slope)
 is removed fits the data as good as one in which this term is present. If it does, it is compared to the random intercept 
 model only.
}
\author{
Federico Rogai
}
