% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{check_slot.study}
\alias{check_slot.study}
\title{Function throws error if specified slot is empty
This function is used in some other functions.}
\usage{
check_slot.study(object, slot)
}
\arguments{
\item{object}{Object of class study}

\item{slot}{Name of slot to check.}
}
\value{
no return value.
}
\description{
Function throws error if specified slot is empty
This function is used in some other functions.
}
\author{
Stefan Thoma
}
