% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{plot_difference.study}
\alias{plot_difference.study}
\title{This function plots the difference table as a forest plot
It calls the function plot.table.study}
\usage{
plot_difference.study(
  object,
  cutoff = NULL,
  ci.type = "wald",
  standardise = TRUE
)
}
\arguments{
\item{object}{a study object}

\item{cutoff}{should plot be cropped?}

\item{ci.type}{either "wald" or "newcombe"
defaults to "wald".}
}
\value{
ggplot
}
\description{
This function plots the difference table as a forest plot
It calls the function plot.table.study
}
\author{
Stefan Thoma
}
