% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{plot_estimates.study}
\alias{plot_estimates.study}
\title{This function plots the relevance_overview_tables}
\usage{
plot_estimates.study(object, cutoff = NULL, standardise = TRUE)
}
\arguments{
\item{object}{a study object}

\item{coverage_probability}{Double. Coverage probability of the CIs. Defaults
to .95


TODO: change measure}
}
\description{
This function plots the relevance_overview_tables
}
\author{
Stefan Thoma
}
