% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{power.f}
\alias{power.f}
\title{Calculates Power given a relvance table object.}
\usage{
power.f(x, std.effect, threshold, family)
}
\arguments{
\item{dat.list}{list containing the full data and other infos from the manyLabs experiment}

\item{coverage_probability}{Double. Coverage probability of the CIs. Defaults
to .95}
}
\description{
Calculates Power given a relvance table object.
}
\author{
Stefan Thoma
}
