% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{r_squared.study}
\alias{r_squared.study}
\title{We want to get the adj. R^2 with bootstrap confidence interval separately for each
lab but also for the MEMo.
We use the package rsq}
\usage{
r_squared.study(object, coverage_prob = 0.95, bootstrap.type = "norm")
}
\arguments{
\item{object}{Study object.}

\item{coverage_prob}{double.}

\item{bootstrap.type}{forwarded to the boot.ci function}
}
\value{
CI. numeric vector containing estimate and CI
}
\description{
We want to get the adj. R^2 with bootstrap confidence interval separately for each
lab but also for the MEMo.
We use the package rsq
}
\author{
Stefan Thoma
}
