% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{relevance_table.study}
\alias{relevance_table.study}
\title{creates a table with the effectsizes of all locations for a specific
 research question. In addition, the function calculates one-sided CIs for
 the effect found by each university. Whether the effect size for a specific
 question is calculated as the standardized mean difference or as the odds
 ratio is determined in the argument "vars_for_question"}
\usage{
relevance_table.study(
  object,
  coverage_prob = 0.95,
  mem = TRUE,
  use.both = TRUE
)
}
\arguments{
\item{object}{study object}

\item{mem}{boolean, whether fixed effect from mixed effects model should be included}

\item{use.both}{boolean, should both data frames be used?}

\item{coverage_probability}{Double. Coverage probability of the CIs. Defaults
to .95}
}
\value{
after_esclalc data.frame. Contains the university-level effect sizes
and CIs for one research question.
}
\description{
creates a table with the effectsizes of all locations for a specific
 research question. In addition, the function calculates one-sided CIs for
 the effect found by each university. Whether the effect size for a specific
 question is calculated as the standardized mean difference or as the odds
 ratio is determined in the argument "vars_for_question"
}
\author{
Stefan Thoma, adapted from Lorenz Herger
}
