% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\docType{data}
\name{scales}
\alias{scales}
\title{Revision protocol dataframe of replication project ML1}
\format{
A data frame with 1149 rows and 4 variables.
}
\source{
\url{}
}
\usage{
scales
}
\description{
The original study conducted by Schwarz (1985) investigated the effect of the design of questionnaire-scales (`scalesgroup`)
on the information reported by participants (un-intuitively labeled `scales`).
Specifically, they randomly assigned $N = 132$ participants either into the low category or the high category group.
Participants were asked to rate how many hours of tv they watched per day on a questionnaire.
The answer options varied depending on the assigned group.
In the low category group six answer categories ranging from *less than 0.5 hours* to *2.5 hours or more* were possible.
In the high category group the six answer categories ranged from *less than 2.5 hours* to *4.5 hours or more*.
They found that in the high category group significantly more people reported watching more than 2.5 hours of tv per day
 compared participants in the the low category group.
`
The Many Labs 1 project conducted replication attempts in 13 locations and
 sample sizes ranged from $N = 71$ to $N = 1261$, totaling $N = 5899$ klein (2014).
We used the schwarz study to illustrate LOR.
}
\details{
See https://osf.io/wx7ck for more infos.
The data has been reduced to only 10 locations.
}
\keyword{datasets}
