% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{stnd.beta.glmer}
\alias{stnd.beta.glmer}
\title{Standardise Effects of glmer-model
This function is used by the relevance_table.study function}
\usage{
stnd.beta.glmer(mod)
}
\arguments{
\item{mod}{any glmer model (probabyl RandomCoefficients Model)}
}
\value{
standardised effects + standardised CIs
}
\description{
Standardise Effects of glmer-model
This function is used by the relevance_table.study function
}
\author{
Stefan Thoma
}
