% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2_study-fn.R
\name{stnd.beta.glmer}
\alias{stnd.beta.glmer}
\title{Standardise Effects of glmer-model
This function is used by the create_overview_table function from replication-fn.R}
\usage{
stnd.beta.glmer(mod)
}
\arguments{
\item{mod}{any glmer model (probabyl RandomCoefficients Model)}
}
\value{
standardised effects + standardised CI based on
}
\description{
Standardise Effects of glmer-model
This function is used by the create_overview_table function from replication-fn.R
}
\author{
Stefan Thoma
}
