% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/study-fn.R
\name{summary.study}
\alias{summary.study}
\title{summary.study, function to summarise the data frames of the study objects.
result depends on the measure OR or SMD defined in the study object.}
\usage{
\method{summary}{study}(object)
}
\arguments{
\item{object}{Object of class study}
}
\value{
some summary measures of the data, mainly about the data
}
\description{
summary.study, function to summarise the data frames of the study objects.
result depends on the measure OR or SMD defined in the study object.
}
\author{
Stefan Thoma
}
